Rename interleaved header
